last changes

+++ src/Symbol.h
@@ -27,7 +27,7 @@ namespace zetscript{
 	struct Symbol {
 		const char *file;   // file where symbol was registered
 		short  line;		// line where symbol was registered
-		short  idx_position;  // used as position where local var/function is located VM (VM reserves space required for all vars)
+		int   idx_position;  // used as position where local var/function is located VM (VM reserves space required for all vars)
 
 		Scope *scope;		// scope where symbol was registered
 		unsigned short properties; // symbol properties using SymbolProperty bits
